requirements_test.txt
-r requirements.txt
pytest-homeassistant-custom-component==0.13.287
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-mock==3.14.0
pytest-timeout==2.4.0
isort==7.0.0
black==25.9.0
pylint>=4.0.1,<5
pre-commit==4.3.0
reorder-python-imports==3.14.0
